:pill: fix: Prevent re-rendering by using React.memo
